Am modificat tipul de la int la byte.

+++ treipigmei-chess-ai/src/board/Board.java
@@ -12,31 +12,31 @@ public class Board {
     public Board() {
         field = new String[8][8];
         
-        setPiece(new int[]{0, 0}, "BR");
-        setPiece(new int[]{0, 1}, "BN");
-        setPiece(new int[]{0, 2}, "BB");
-        setPiece(new int[]{0, 3}, "BQ");
-        setPiece(new int[]{0, 4}, "BK");
-        setPiece(new int[]{0, 5}, "BB");
-        setPiece(new int[]{0, 6}, "BN");
-        setPiece(new int[]{0, 7}, "BR");
+        setPiece(new byte[]{0, 0}, "BR");
+        setPiece(new byte[]{0, 1}, "BN");
+        setPiece(new byte[]{0, 2}, "BB");
+        setPiece(new byte[]{0, 3}, "BQ");
+        setPiece(new byte[]{0, 4}, "BK");
+        setPiece(new byte[]{0, 5}, "BB");
+        setPiece(new byte[]{0, 6}, "BN");
+        setPiece(new byte[]{0, 7}, "BR");
         
-        for (int i = 0; i < 8; i++) {
-            setPiece(new int[]{1, i}, "BP");
-            for (int j = 2; j <= 5; j++) {
-                setPiece(new int[]{j, i}, "##");
+        for (byte i = 0; i < 8; i++) {
+            setPiece(new byte[]{1, i}, "BP");
+            for (byte j = 2; j <= 5; j++) {
+                setPiece(new byte[]{j, i}, "##");
             }
-            setPiece(new int[]{6, i}, "WP");
+            setPiece(new byte[]{6, i}, "WP");
         }
         
-        setPiece(new int[]{7, 0}, "WR");
-        setPiece(new int[]{7, 1}, "WN");
-        setPiece(new int[]{7, 2}, "WB");
-        setPiece(new int[]{7, 3}, "WQ");
-        setPiece(new int[]{7, 4}, "WK");
-        setPiece(new int[]{7, 5}, "WB");
-        setPiece(new int[]{7, 6}, "WN");
-        setPiece(new int[]{7, 7}, "WR");
+        setPiece(new byte[]{7, 0}, "WR");
+        setPiece(new byte[]{7, 1}, "WN");
+        setPiece(new byte[]{7, 2}, "WB");
+        setPiece(new byte[]{7, 3}, "WQ");
+        setPiece(new byte[]{7, 4}, "WK");
+        setPiece(new byte[]{7, 5}, "WB");
+        setPiece(new byte[]{7, 6}, "WN");
+        setPiece(new byte[]{7, 7}, "WR");
     }
     
     /**
@@ -44,9 +44,9 @@ public class Board {
      * @param pos
      * @param piece
      */
-    public void setPiece(int [] pos, String piece) {
-        int i = pos[0];
-        int j = pos[1];
+    public void setPiece(byte [] pos, String piece) {
+        byte i = pos[0];
+        byte j = pos[1];
         field[i][j] = piece;
     }
     
@@ -55,9 +55,9 @@ public class Board {
      * @param pos
      * @return
      */
-    public String getPiece(int [] pos) {
-        int i = pos[0];
-        int j = pos[1];
+    public String getPiece(byte [] pos) {
+        byte i = pos[0];
+        byte j = pos[1];
         return field[i][j];
     }
